+++ SeriesGuide/src/androidTest/java/com/battlelancer/seriesguide/test/EpisodeToolsTest.java
@@ -1,29 +1,32 @@
-package com.battlelancer.seriesguide.util;
+package com.battlelancer.seriesguide.test;
 
 import com.battlelancer.seriesguide.enums.EpisodeFlags;
-
-import org.junit.Test;
+import com.battlelancer.seriesguide.util.EpisodeTools;
+import junit.framework.Test;
+import junit.framework.TestCase;
+import junit.framework.TestSuite;
 
 import static org.fest.assertions.api.Assertions.assertThat;
 import static org.fest.assertions.api.Assertions.failBecauseExceptionWasNotThrown;
 
-public class EpisodeToolsTest {
+public class EpisodeToolsTest extends TestCase {
 
-    @Test
+    public static Test suite() {
+        return new TestSuite(EpisodeToolsTest.class);
+    }
+
     public void test_isWatched() {
         assertThat(EpisodeTools.isWatched(EpisodeFlags.WATCHED)).isTrue();
         assertThat(EpisodeTools.isWatched(EpisodeFlags.SKIPPED)).isFalse();
         assertThat(EpisodeTools.isWatched(EpisodeFlags.UNWATCHED)).isFalse();
     }
 
-    @Test
     public void test_isUnwatched() {
         assertThat(EpisodeTools.isUnwatched(EpisodeFlags.UNWATCHED)).isTrue();
         assertThat(EpisodeTools.isUnwatched(EpisodeFlags.WATCHED)).isFalse();
         assertThat(EpisodeTools.isUnwatched(EpisodeFlags.SKIPPED)).isFalse();
     }
 
-    @Test
     public void test_validateFlags() {
         EpisodeTools.validateFlags(EpisodeFlags.UNWATCHED);
         EpisodeTools.validateFlags(EpisodeFlags.WATCHED);