+++ SeriesGuide/src/main/java/com/battlelancer/seriesguide/ui/EpisodeDetailsFragment.java
@@ -369,7 +369,7 @@ public class EpisodeDetailsFragment extends SherlockFragment implements ActionsF
             // "in 15 mins (Fri)"
             timeAndNumbersText
                     .append(getString(R.string.release_date_and_day,
-                            TimeTools.formatToRelativeLocalReleaseTime(actualRelease),
+                            TimeTools.formatToRelativeLocalReleaseTime(getActivity(), actualRelease),
                             TimeTools.formatToLocalReleaseDay(actualRelease))
                             .toUpperCase(Locale.getDefault()));
             timeAndNumbersText.append("  ");